requirements.txt
rembg
onnxruntime
easyocr
paddleocr
paddlepaddle
numpy
pillow
scikit-image
rembg
opencv-python
flask